맵 css 수정

+++ front/src/components/board/BoardDetail.vue
@@ -134,21 +134,22 @@ export default {
       }
     },
     heartClick() {
-      if (this.heartChk === 0) {
-        http.post(`boardlike/${this.board.boardNo}/${this.userNo}`).then((response) => {
-          ++this.heartCnt;
-          console.log(response);
-        });
-        this.heart = require("@/assets/heartOn.png");
-      } else {
-        http.delete(`boardlike/${this.board.boardNo}/${this.userNo}`).then((response) => {
-          --this.heartCnt;
-          console.log(response);
-        });
-        this.heart = require("@/assets/heartOff.png");
+      if (this.userInfo !== null) {
+        if (this.heartChk === 0) {
+          http.post(`boardlike/${this.board.boardNo}/${this.userNo}`).then((response) => {
+            ++this.heartCnt;
+            console.log(response);
+          });
+          this.heart = require("@/assets/heartOn.png");
+        } else {
+          http.delete(`boardlike/${this.board.boardNo}/${this.userNo}`).then((response) => {
+            --this.heartCnt;
+            console.log(response);
+          });
+          this.heart = require("@/assets/heartOff.png");
+        }
+        this.heartChk = this.heartChk === 0 ? 1 : 0;
       }
-      this.heartChk = this.heartChk === 0 ? 1 : 0;
-      this.$forceUpdate();
     },
     postComment() {
       let commentData = {
